# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VeriFace Edge API",
    "version": "1.0.0",
    "description": "Privacy-first web facial authentication SDK. All biometric computation runs in the browser via WebGPU/WASM. Backend receives only zero-knowledge Pedersen commitments + signed JWTs.",
    "contact": {
      "name": "VeriFace Edge Support",
      "email": "support@veriface.io"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "https://api.veriface.io", "description": "Production" },
    { "url": "https://staging.api.veriface.io", "description": "Staging" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Format: `Bearer vf_live_...` or `Bearer vf_test_...`"
      }
    },
...
  },
  "paths": {
...
    "/api/webauthn/register/begin": {
      "post": {
        "summary": "Begin FIDO2/WebAuthn credential registration",
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "externalUserId": { "type": "string" },
                  "deviceType": { "type": "string", "enum": ["platform", "roaming"] }
                },
                "required": ["externalUserId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "PublicKeyCredentialCreationOptions" }
        }
      }
    },
    "/api/webauthn/register/finish": {
      "post": {
        "summary": "Complete WebAuthn credential registration",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": { "description": "Credential stored" }
        }
      }
    },
...
    "/api/webauthn/auth/finish": {
      "post": {
        "summary": "Complete WebAuthn authentication",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": { "description": "Assertion verified" }
        }
      }
    },
...
  }
}